refactor(sx.js): move Starknet types to starknet clients
We used to store EVM related types in clients/evm/types, but
Starknet related types were stuck in global types.

To unify it we now moved those Starknet specific types to
clients/starknet/types.

This is type only change and shouldn't affect any logic.
