+++ packages/sx.js/src/authenticators/starknet/starkSig.ts
@@ -9,7 +9,7 @@ import {
   UpdateProposalCallArgs,
   Vote,
   VoteCallArgs
-} from '../../types';
+} from '../../clients/starknet/types';
 import { getChoiceEnum } from '../../utils/starknet-enums';
 
 const callData = new CallData(StarkSigAuthenticatorAbi);
