+++ packages/sx.js/src/strategies/starknet/index.ts
@@ -3,7 +3,8 @@ import createEvmSlotValueStrategy from './evmSlotValue';
 import createMerkleWhitelistStrategy from './merkleWhitelist';
 import createOzVotesStorageProofStrategy from './ozVotesStorageProof';
 import createVanillaStrategy from './vanilla';
-import { NetworkConfig, Strategy } from '../../types';
+import { Strategy } from '../../clients/starknet/types';
+import { NetworkConfig } from '../../types';
 import { hexPadLeft } from '../../utils/encoding';
 
 export function getStrategy(
